Checkpoint
Seperate directory for Admin UI and Client UI
Creating service layer in all Implementations

+++ W7Project/Project/W7Project/src/main/java/com/oocl/mnlbc/dao/GenericDAO.java
@@ -15,5 +15,5 @@ public interface GenericDAO<T extends Serializable> {
 
 	void delete(final T entity);
 
-	void deleteById(final long entityId);
+	void deleteById(final int entityId);
 }
